+++ public/app/style.css
@@ -85,4 +85,14 @@ h1 {
 
 .reg-text {
     text-align: left;
+    color: white;
+    font-size: 35px; 
+    font-weight: bolder;
+}
+
+.form-text {
+    box-shadow: inset 0 0 0 1000px rgba(0, 0, 0, 0.098);
+    background-color: rgba(255, 255, 255, 0.111);
+    text-align: left;
+    color: white;
 }
